Pickling/requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
MarkupSafe==1.0

numpy
flask-cors
flask_restplus
scikit-learn
keras
tensorflow
pandas
python-dotenv
